Fix dense_evolution.qft/random_circuit module-name-shadowing bug
Both the top-level package and a flat backward-compat shim file used
the same attribute name (qft/random_circuit). Python's import machinery
unconditionally sets a submodule as an attribute on its parent package
the moment it's imported, so any code importing the flat shim path
clobbered the re-exported function into a module -- reproducible as
TypeError: 'module' object is not callable whenever test_qft.py/
test_random_circuit.py ran alongside test_mps.py's dashboard-engine
tests (qasm_library.py calls de.qft(3)/de.random_circuit(...) at
import time).

qft had already moved to circuits/qft.py; random_circuit now follows
the same pattern (real implementation in circuits/random_circuit.py,
flat random_circuit.py reduced to a shim), and both test files import
from the real subpackage path instead of the flat shim.

Residual risk, not eliminated by design: external code that still
imports the flat shim path directly triggers the same clobbering.
