+++ README.md
@@ -857,6 +857,7 @@ All circuits stored as OpenQASM 2.0 strings in `dashboard_core.QASM_LIBRARY`.
 ## ▍ Changelog
 
 ### Unreleased
+- **Fixed a real module/function name collision: `dense_evolution.qft`/`dense_evolution.random_circuit` could silently turn into an unusable module.** Both the top-level package (`from .circuits.qft import qft`) and a flat backward-compat shim file (`dense_evolution/qft.py`) used the exact same attribute name, `qft` -- Python's import machinery unconditionally sets a submodule as an attribute on its parent package the moment it's imported, regardless of what `__init__.py` already bound there. Any code that imported the flat shim path directly (`from dense_evolution.qft import qft`, or the equivalent `random_circuit` case) silently clobbered `dense_evolution.qft` from the callable function to the module object -- every later `dense_evolution.qft(...)` call in the same process then failed with `TypeError: 'module' object is not callable`. This was live and reproducible in this repo's own test suite (`tests/unit/test_qft.py`/`test_random_circuit.py` imported the flat shim, `tools/dashboard_core/qasm_library.py` called `de.qft(3)`/`de.random_circuit(...)` at import time -- three `test_mps.py` dashboard-engine tests failed with this exact error whenever both test files ran in the same pytest session), not merely a theoretical risk. `qft` had already moved to `circuits/qft.py`, so the fix moves `random_circuit` the same way (real implementation now `dense_evolution/circuits/random_circuit.py`, flat `dense_evolution/random_circuit.py` reduced to a shim), and both test files now import from the real subpackage path instead of the flat shim. **Residual risk, not eliminated by design**: any *external* code (e.g. Dense-Evolution-Discovery) that still imports the flat shim path directly triggers the same clobbering -- verified directly (`from dense_evolution.qft import qft` after `import dense_evolution` does still replace the attribute); the flat paths are kept only for backward compatibility and were never safe to import as submodules in the same process as other `dense_evolution.qft`/`.random_circuit` usage.
 - **Removed the fictitious CuPy/CUDA GPU path.** `DenseSVSimulator`/`Chunk`'s `use_gpu` parameter was accepted and forwarded everywhere but never read -- its own docstring said "reserved for future CuPy/JAX GPU dispatch", and a full-repo audit found zero `cp.<anything>()` calls: `circuits/registry.py` imported `cupy` only to set an unused `has_cupy` diagnostics flag. The README's own feature table claimed "CuPy CUDA -- runtime selection, zero code changes", which was false. GPU acceleration in this library has only ever been JAX's own automatic device placement (`self.xp = jnp`, dispatched to whatever device `jax.devices()` reports) -- real, but unrelated to CuPy or the `use_gpu` flag. Fixed: `use_gpu` removed from `DenseSVSimulator`/`Chunk` (a no-op parameter, so nothing could have depended on its behavior -- only its presence in the signature, if passed explicitly by keyword); the unused `cupy` import/`has_cupy` flag removed from `circuits/registry.py`; the `[gpu]` extra now installs `jax[cuda12]` (the package that actually matters) instead of `cupy-cuda12x`; README/`pyproject.toml` corrected to describe the real JAX-based GPU path.
 ### v8.1.62
 - **New `mitigation.sample_classical_shadow`/`magic_entropy_from_shadows`/`approx_shadow_std`/`fit_shadow_sample_complexity`: a classical-shadows-based estimator for `magic_entropy` from randomized measurement snapshots.** Distinct API shape from every other function in `mitigation` -- sampling (`sample_classical_shadow`) and estimation (`magic_entropy_from_shadows`) are separate steps, since shadow data can come from this simulator (oracle Born-rule sampling) or, in principle, real hardware measurement logs reconstructed the same way. Originated from a Colab proposal for a `shadows.py` module with a real bug (a missing transpose in its purity estimator's U-statistic contraction, silent whenever every snapshot happened to be real-valued) -- fixed, then extended to `magic_entropy` using the same multi-copy trick Huang, Kueng & Preskill (2020, [arXiv:2002.08953](https://arxiv.org/abs/2002.08953)) state "readily generalizes to higher order polynomials." Matured across three real gaps found and closed in [Dense-Evolution-Discovery](https://github.com/tatopenn-cell/Dense-Evolution-Discovery) Experiment 31 before promotion: the purity-estimator bug itself, plain averaging replaced with real median-of-means robustness (verified: tolerates a 40%-corrupted contiguous measurement block, unlike a naive mean, which is dragged from 1.0 to -19.4 under the same corruption), and an empirically-fitted sample-complexity curve (`std(n) ~ 11.75 / n^0.546` from 20 independent trials per snapshot count, exponent consistent with the ~0.5 theory predicts) exposed as `approx_shadow_std` (a rough built-in guide) and `fit_shadow_sample_complexity` (calibrate your own curve for a specific state). Not `jax.jit`-compatible, unlike the rest of this subpackage -- median-of-means uses `numpy.median`, which has no equivalent JAX primitive at this scale.
