+++ dense_evolution/circuits/__init__.py
@@ -5,6 +5,7 @@ from .compiler import QuantumTranspiler
 from .registry import HAS_JAX, NoiseModel, NoiseSpec, QuantumHardwareRegistry
 from .topology import entangling_layer, VALID_PATTERNS
 from .qft import qft
+from .random_circuit import random_circuit
 from .trotter import pauli_rotation_ops, trotter_evolve_ops
 
 __all__ = [
@@ -14,5 +15,6 @@ __all__ = [
     "HAS_JAX", "NoiseModel", "NoiseSpec", "QuantumHardwareRegistry",
     "entangling_layer", "VALID_PATTERNS",
     "qft",
+    "random_circuit",
     "pauli_rotation_ops", "trotter_evolve_ops",
 ]
